tests/wpt/tests/html/semantics/embedded-content/media-elements/no-autoplay-audio-history-back-does-not-play.html

<!doctype html>
<meta charset="utf-8">
<title>No autoplay audio does not start after history back</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
  <script>
    "use strict";

    function runHarness() {
      promise_test(async t => {
        const base = location.pathname.replace(/[^/]*$/, "");

        const iframe = document.createElement("iframe");
        document.body.appendChild(iframe);
        t.add_cleanup(() => iframe.remove());

        function waitForMessage(type, timeoutMs) {
          return new Promise((resolve, reject) => {
            const timeout = step_timeout(() => {
              window.removeEventListener("message", onMessage);
              reject(new Error(`Timed out waiting for ${type}`));
            }, timeoutMs);

            function onMessage(event) {
              if (event.source !== iframe.contentWindow) return;
              if (!event.data || event.data.type !== type) return;
              clearTimeout(timeout);
              window.removeEventListener("message", onMessage);
              resolve(event.data);
            }

            window.addEventListener("message", onMessage);
          });
        }

        iframe.src = `${base}resources/no-autoplay-audio-history-back-does-not-play-a.html`;
        await waitForMessage("ready_a", 5000);


        iframe.src = `${base}resources/no-autoplay-audio-history-back-does-not-play-b.html`;
        await waitForMessage("loaded_b", 5000);


        iframe.contentWindow.history.back();


        const result = await waitForMessage("result", 5000);

        assert_true(result.pass, JSON.stringify(result.details));
      }, "No autoplay audio remains paused after iframe history back");
    }

    runHarness();
  </script>
</body>
